add macros

+++ FitnessServices/Models/Macros.cs
@@ -2,11 +2,11 @@ namespace FitnessServices.Models;
 
 public sealed class Macros
 {
-    public int Calories { get; set; }
-    public int Protein { get; set; }
-    public int Fat { get; set; }
-    public int Carbs { get; set; }
-    public int Fiber { get; set; }
-    public int Alcohol { get; set; }
-    public int Water { get; set; }
+    public float Calories { get; set; }
+    public float Protein { get; set; }
+    public float Fat { get; set; }
+    public float Carbs { get; set; }
+    public float Fiber { get; set; }
+    public float Alcohol { get; set; }
+    public float Water { get; set; }
 }
